MVC 모델 적용 (1차)
