+++ src/views/startView.js
@@ -1,0 +1,7 @@
+const { Console } = require('@woowacourse/mission-utils');
+
+function startView(){
+    Console.print('숫자 야구 게임을 시작합니다.');
+};
+
+exports.startView = startView;
